Scope test data to functions
As the test suite is growing, test data at the time loses context. Scope
the fixture variables to be within the relevent test function.

This commit also removes the variables holding time values, as we can
simply define them with the aws helper methods to create pointers to
time.Time types.